refactor: divide component 'BrownCard' from the code of question
In tutorial page, the html div which class name is 'questionA' and 'questionB' contain
these elements: question-text, identify-box for display satellite images, button and the answer card.
However, when tutorial page finished, these element excepted the answer card will be used angin, which means the component 'card' is not reusable.
So I decide to seperate the card from the code of question in HTML, making the component of the question can reuse in after pages.
Note: the component of qusetion hasn't completed, I will update it ASAP.