refactor: change the code of the question into component
In order to increase the readability and reusability of the code, I reconstruct the code of the tutorial page,
changing the code of the question into two components, "TutorialQuestionA"  and "TutorialQuestionB".
Compared to older version, the style and function of tutorial can be almost accomplished by three components:
"TutorialQuestionA", "TutorialQuestionB", "BrownCard".
However, these components still need further revise if we want to use these components not only in tutorial but in the game afterwards.